adjusting the test routine for saving results

+++ LeSpMV/include/spmv_testroutine.h
@@ -11,10 +11,10 @@ template <typename IndexType, typename ValueType>
 double test_csr_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 template <typename IndexType, typename ValueType>
-int test_bsr_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
+double test_bsr_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 template <typename IndexType, typename UIndexType, typename ValueType>
-int test_csr5_matrix_kernels(const CSR_Matrix<IndexType, ValueType> &csr_ref, int kernel_tag, int schedule_mod);
+double test_csr5_matrix_kernels(const CSR_Matrix<IndexType, ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 /**
  * @brief Input CSR format for reference. Inside we make an ELL matrix
@@ -23,13 +23,13 @@ int test_csr5_matrix_kernels(const CSR_Matrix<IndexType, ValueType> &csr_ref, in
  * @tparam ValueType 
  * @param csr 
  * @param kernel_tag 
- * @return int 
+ * @return double time in ms 
  */
 template <typename IndexType, typename ValueType>
-int test_ell_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, LeadingDimension ld, int schedule_mod);
+double test_ell_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, LeadingDimension ld, int schedule_mod);
 
 template <typename IndexType, typename ValueType>
-int test_dia_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
+double test_dia_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 /**
  * @brief Input CSR format for reference. Inside we make an S_ELL matrix for testing
@@ -41,12 +41,12 @@ int test_dia_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int 
  * @return int 
  */
 template <typename IndexType, typename ValueType>
-int test_s_ell_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
+double test_s_ell_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 template <typename IndexType, typename ValueType>
-int test_sell_c_sigma_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
+double test_sell_c_sigma_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 template <typename IndexType, typename ValueType>
-int test_sell_c_R_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
+double test_sell_c_R_matrix_kernels(const CSR_Matrix<IndexType,ValueType> &csr_ref, int kernel_tag, int schedule_mod);
 
 #endif /* SPMV_TESTROUTINE_H */
